pyproject.toml
[project]
name = "web3py-favorites"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv>=1.0.1",
    "vyper>=0.4.0",
    "web3>=7.6.0",
]
